Fix delete for shutdown experiments
Delete experiment would fail if the experiment did not start
at all.

+++ floyd/cli/experiment.py
@@ -165,7 +165,8 @@ def delete(id, yes):
     Delete project run
     """
     experiment = ExperimentClient().get(id)
-    task_instance = TaskInstanceClient().get(get_module_task_instance_id(experiment.task_instances))
+    task_instance_id = get_module_task_instance_id(experiment.task_instances)
+    task_instance = TaskInstanceClient().get(task_instance_id) if task_instance_id else None
 
     if experiment.state in ["queued", "running"]:
         floyd_logger.info("Experiment in {} state cannot be deleted. Stop it first".format(experiment.state))
@@ -174,7 +175,7 @@ def delete(id, yes):
     if not yes:
         click.confirm('Delete Run: {}?'.format(experiment.name), abort=True, default=False)
 
-    if task_instance.module_id:
+    if task_instance and task_instance.module_id:
         ModuleClient().delete(task_instance.module_id)
 
     if ExperimentClient().delete(id):
